Refactored packagename, changed name to query in POJO class.
