+++ app/src/main/java/fr/free/nrw/commons/data/DBOpenHelper.java
@@ -6,7 +6,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 
 import fr.free.nrw.commons.category.CategoryDao;
 import fr.free.nrw.commons.contributions.ContributionDao;
-import fr.free.nrw.commons.explore.recent_searches.RecentSearchesDao;
+import fr.free.nrw.commons.explore.recentsearches.RecentSearchesDao;
 import fr.free.nrw.commons.modifications.ModifierSequenceDao;
 
 public class DBOpenHelper  extends SQLiteOpenHelper {
