+++ app/src/main/java/fr/free/nrw/commons/di/CommonsApplicationModule.java
@@ -21,7 +21,7 @@ import fr.free.nrw.commons.upload.UploadController;
 
 import static android.content.Context.MODE_PRIVATE;
 import static fr.free.nrw.commons.contributions.ContributionsContentProvider.CONTRIBUTION_AUTHORITY;
-import static fr.free.nrw.commons.explore.recent_searches.RecentSearchesContentProvider.RECENT_SEARCH_AUTHORITY;
+import static fr.free.nrw.commons.explore.recentsearches.RecentSearchesContentProvider.RECENT_SEARCH_AUTHORITY;
 import static fr.free.nrw.commons.modifications.ModificationsContentProvider.MODIFICATIONS_AUTHORITY;
 
 @Module
