+++ app/src/main/java/fr/free/nrw/commons/di/ContentProviderBuilderModule.java
@@ -4,7 +4,7 @@ import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
 import fr.free.nrw.commons.category.CategoryContentProvider;
 import fr.free.nrw.commons.contributions.ContributionsContentProvider;
-import fr.free.nrw.commons.explore.recent_searches.RecentSearchesContentProvider;
+import fr.free.nrw.commons.explore.recentsearches.RecentSearchesContentProvider;
 import fr.free.nrw.commons.modifications.ModificationsContentProvider;
 
 @Module
