+++ app/src/main/java/fr/free/nrw/commons/explore/SearchActivity.java
@@ -20,7 +20,7 @@ import butterknife.ButterKnife;
 import fr.free.nrw.commons.Media;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.explore.images.SearchImageFragment;
-import fr.free.nrw.commons.explore.recent_searches.RecentSearchesFragment;
+import fr.free.nrw.commons.explore.recentsearches.RecentSearchesFragment;
 import fr.free.nrw.commons.media.MediaDetailPagerFragment;
 import fr.free.nrw.commons.theme.NavigationBaseActivity;
 import fr.free.nrw.commons.utils.ViewUtil;
