+++ app/src/main/res/layout/fragment_search_history.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="fr.free.nrw.commons.explore.recent_searches.RecentSearchesFragment">
+    tools:context="fr.free.nrw.commons.explore.recentsearches.RecentSearchesFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
