+++ app/src/main/res/values/strings.xml
@@ -291,6 +291,4 @@
     <string name="wallpaper_set_successfully">Wallpaper set successfully!</string>
     <string name="search_recent_header">Recent searches:</string>
     <string name="provider_searches">Recently searched queries</string>
-    <string name="delete_recent_searches_dialog">Are you sure you want to clear your search history?</string>
-    <string name="search_history_deleted">Search history deleted</string>
 </resources>
